fix: resolve 6 blocking issues from devil's advocate review
C1 (CRITICAL): Standardize embedding model name to clap-htsat-large
C2 (CRITICAL): Sanitize database credentials in error messages
C3 (CRITICAL): Convert IngestResponse.status to StrEnum
H4 (HIGH): Change file_size_bytes from Integer to BigInteger
H5 (HIGH): Add pyright to dev dependencies
M7 (MEDIUM): Fix error code INTERNAL_SERVER_ERROR -> INTERNAL_ERROR

All 3 tests pass, pyright clean, no regressions.

+++ audio-ident-service/app/models/track.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, Float, Index, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, Float, Index, Integer, String, Text, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -27,7 +27,7 @@ class Track(Base):
 
     # File identity
     file_hash_sha256: Mapped[str] = mapped_column(String(64), unique=True, nullable=False)
-    file_size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
+    file_size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
     file_path: Mapped[str] = mapped_column(Text, nullable=False)
 
     # Chromaprint (ingestion-time content dedup ONLY)
